+++ Src/Rect2D.fs
@@ -66,7 +66,7 @@ type Rect2D =
     /// Create a 2D rectangle from origin coordinates and X- and Y-axis vector components.
     [<Obsolete("Unsafe internal constructor, doesn't check the input (unless compiled in DEBUG mode), but must be public for inlining. So marked Obsolete instead.") >]
     new (originX:float, originY:float, axisXX:float, axisXY:float, axisYX:float, axisYY:float) =
-        #if DEBUG || CHECK_EUCLID // CHECK_EUCLID so checks can still be enabled when using with Fable release mode
+        #if DEBUG || CHECKED_EUCLID // CHECKED_EUCLID so checks can still be enabled when using with Fable release mode
             let lenX = sqrt(axisXX*axisXX + axisXY*axisXY)
             let lenY = sqrt(axisYX*axisYX + axisYY*axisYY)
             if isTooSmall lenX then  failTooSmall2 "Rect2D() axisX" (Vc(axisXX, axisXY)) (Vc(axisYX, axisYY))